Extract TeachingFabricService from FabricService
Move lesson-only canvas operations (zones, evaluation, feedback) and
their private helpers into a new TeachingFabricService under
playground/teaching/. FabricService is now a pure canvas wrapper;
TeachingPlaygroundService injects the new service for lesson flow.
